Introduce `forceResolveLazyTypes` into TypeResolutionContext
The main reason is to force type resolution when resolving
type arguments in class literal as annotation value argument:

@Ann(Array<String>::class) class Ann

`String` left unresolved unless this commit is applied.

Note, that currently forced calculations used only when bare types are allowed
as it's true for class literals resolution and currenlty it's enough.

GitOrigin-RevId: 24f56d3ee10d2e66fc207c0d5c1ad0a3c08ef42a